+++ src/logic/IDGenerator.h
@@ -5,7 +5,7 @@
 #ifndef TODOLIST_IDGENERATOR_H
 #define TODOLIST_IDGENERATOR_H
 
-#include "model/TaskID.h"
+#include "api/TaskID.h"
 
 class IDGenerator {
