+++ tests/model/TaskIDTest.cpp
@@ -4,7 +4,7 @@
 
 
 #include <gtest/gtest.h>
-#include "model/TaskID.h"
+#include "api/TaskID.h"
 
 class TaskIDTest : public ::testing::Test {
